requirements.txt
streamlit
pdfplumber
langchain
faiss-cpu
sentence-transformers
google-generativeai
